--- src/components/FormReadyHelp/DateOfBirth/DateOfBirth.module.scss
@@ -1,7 +1,0 @@
-@import '../../../styles/index.scss';
-
-.dob {
-  &__container {
-    grid-area: dob;
-  }
-}
