--- src/components/FormReadyHelp/FirstName/FirstName.module.scss
@@ -1,5 +1,0 @@
-.firstName {
-  &__container {
-    grid-area: firstName;
-  }
-}
