--- src/components/FormReadyHelp/LastName/LastName.module.scss
@@ -1,5 +1,0 @@
-.lastName {
-  &__container {
-    grid-area: lastName;
-  }
-}
