--- src/components/FormReadyHelp/SelectCountry/SelectCountry.module.scss
@@ -1,11 +1,0 @@
-@import '../../../styles/index.scss';
-
-.country {
-  &__input {
-    width: 24rem;
-
-    @media (max-width: $tablet) {
-      width: 22.875rem;
-    }
-  }
-}
