--- src/components/FormReadyHelp/Skills/Skills.module.scss
@@ -1,15 +1,0 @@
-@import '../../../styles/index.scss';
-
-.skills {
-  &__input {
-    width: 49.563rem;
-
-    @media (max-width: $tablet) {
-      width: 46.25rem;
-    }
-
-    @media (max-width: $mobile) {
-      width: 22.875rem;
-    }
-  }
-}
