+++ src/components/form/Checkboxes/Checkboxes.module.scss
@@ -1,0 +1,7 @@
+.checkboxes {
+  grid-area: checkboxes;
+
+  &__item {
+    font-weight: 400;
+  }
+}
