+++ src/components/form/Comment/Comment.module.scss
@@ -1,0 +1,14 @@
+@import '../../../styles/index.scss';
+
+.comment {
+  grid-area: comment;
+  width: 49.563rem;
+
+  @media (max-width: $tablet) {
+    width: 46.25rem;
+  }
+
+  @media (max-width: $mobile) {
+    width: 22.875rem;
+  }
+}
