+++ src/components/form/FirstName/FirstName.module.scss
@@ -1,0 +1,3 @@
+.firstName {
+  grid-area: firstName;
+}
