+++ src/components/form/FormReadyHelp/DateOfBirth/DateOfBirth.module.scss
@@ -1,0 +1,7 @@
+@import '../../../../styles/index.scss';
+
+.dob {
+  &__container {
+    grid-area: dob;
+  }
+}
