+++ src/components/form/FormReadyHelp/Language/Language.module.scss
@@ -1,0 +1,36 @@
+@import '../../../../styles/index.scss';
+
+.languages {
+  grid-area: languages;
+}
+
+.language {
+  display: flex;
+  align-items: center;
+  gap: 1.5rem;
+
+  @media (max-width: $tablet) {
+    gap: 0.5rem;
+  }
+
+  @media (max-width: $mobile) {
+    flex-direction: column;
+  }
+
+  &__input {
+    font-weight: 400;
+    width: 24.063rem;
+
+    @media (max-width: $tablet) {
+      width: 22.875rem;
+    }
+  }
+
+  &__button {
+    &_add {
+      display: flex;
+      align-items: center;
+      gap: 0.563rem;
+    }
+  }
+}
