+++ src/components/form/FormReadyHelp/SelectCountry/SelectCountry.module.scss
@@ -1,0 +1,14 @@
+@import '../../../../styles/index.scss';
+
+.country {
+  grid-area: country;
+
+  &__input {
+    font-weight: 400;
+    width: 24rem;
+
+    @media (max-width: $tablet) {
+      width: 22.875rem;
+    }
+  }
+}
