+++ src/components/form/FormReadyHelp/Skills/Skills.module.scss
@@ -1,0 +1,18 @@
+@import '../../../../styles/index.scss';
+
+.skills {
+  grid-area: skills;
+
+  &__input {
+    font-weight: 400;
+    width: 49.563rem;
+
+    @media (max-width: $tablet) {
+      width: 46.25rem;
+    }
+
+    @media (max-width: $mobile) {
+      width: 22.875rem;
+    }
+  }
+}
