+++ src/components/form/FormReadyHelp/Skills/Skills.tsx
@@ -1,0 +1,46 @@
+import { useFormContext, Controller } from 'react-hook-form';
+import Select from 'react-select';
+import { skillsList } from '../../../../utils/skillsList';
+import FormItemWrapper from '../../FormItemWrapper/FormItemWrapper';
+// import FormItemContainer from '../../FormItemContainer';
+// import FormLabel from '../../FormLabel';
+import customStyles from '../../../UI/Form/SelectInput/selectStyle';
+import s from './Skills.module.scss';
+
+const Skills = () => {
+  const { control } = useFormContext();
+
+  const skills = skillsList.map((item) => ({
+    value: item.skill,
+    label: item.skill,
+  }));
+
+  return (
+    <FormItemWrapper
+      className={s.skills}
+      title='Choose your skills (you can select several options) *'
+    >
+      <Controller
+        name='skills'
+        control={control}
+        defaultValue={null}
+        render={({ field }) => (
+          <Select
+            {...field}
+            options={skills}
+            placeholder='Search your skills'
+            styles={customStyles}
+            isMulti
+            onChange={(selectedOption) => {
+              field.onChange(selectedOption);
+            }}
+            className={s.skills__input}
+          />
+        )}
+        rules={{ required: true }}
+      />
+    </FormItemWrapper>
+  );
+};
+
+export default Skills;
