+++ src/components/form/FormReadyHelp/SocialNetwork/SocialNetwork.module.scss
@@ -1,0 +1,30 @@
+@import '../../../../styles/index.scss';
+
+.networks {
+  grid-area: networks;
+
+  &__wrap {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    gap: 1.438rem;
+
+    @media (max-width: $tablet) {
+      gap: 0.25rem;
+    }
+  }
+
+  &__input {
+    &_logo {
+      width: 4.875rem;
+      text-align: center;
+      & > div {
+        padding: 11px 0 !important;
+      }
+    }
+
+    &_network {
+      width: 17.625rem;
+    }
+  }
+}
