+++ src/components/form/LastName/LastName.module.scss
@@ -1,0 +1,3 @@
+.lastName {
+  grid-area: lastName;
+}
